+++ firmware/controllers/actuators/idle_hardware.cpp
@@ -148,7 +148,6 @@ void initIdleHardware() {
 		 */
 		// todo: even for double-solenoid mode we can probably use same single SimplePWM
 		startSimplePwm(&idleSolenoidOpen, "Idle Valve Open",
-			&engine->scheduler,
 			&enginePins.idleSolenoidPin,
 			engineConfiguration->idle.solenoidFrequency, PERCENT_TO_DUTY(engineConfiguration->manIdlePosition));
 
@@ -159,7 +158,6 @@ void initIdleHardware() {
 			}
 
 			startSimplePwm(&idleSolenoidClose, "Idle Valve Close",
-				&engine->scheduler,
 				&enginePins.secondIdleSolenoidPin,
 				engineConfiguration->idle.solenoidFrequency, PERCENT_TO_DUTY(engineConfiguration->manIdlePosition));
 		}