+++ firmware/CHANGELOG.md
@@ -30,6 +30,12 @@ or
 
 ## Unreleased
 
+### Added
+ - Allow fractional tachometer pulse ratio for fine tachometer calibration
+ - Add an option to ramp the idle target down as engine speed returns to idle. Makes the running -> idle transition much smoother while in closed loop mode #570
+
+## February 2025 Release
+
 ### Breaking Changes
  - Large data logging improvements and changes - in TunerStudio, you may need to re-create your dash layouts from default and then re-add any broken gauges and indicators.